% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_Avg.R
\name{agg_Avg}
\alias{agg_Avg}
\title{Aggregated Knockoff with Average Test Statistics}
\usage{
agg_Avg(Ws_mat, fdr = 0.05, offset = 0)
}
\arguments{
\item{Ws_mat}{A matrix where each row represents the test statistics from a different knockoff filter and each column corresponds to a variable.}

\item{fdr}{A numeric value specifying the target false discovery rate (FDR) level. Default is \eqn{0.05}.}

\item{offset}{An integer (0 or 1) specifying the offset in the threshold calculation. Default is \eqn{0} (liberal control).}
}
\value{
A vector \code{shat} containing the indices of selected variables after averaging the test statistics and applying the knockoff filter.
}
\description{
Performs Aggregated Knockoff analysis by averaging the test statistics across multiple knockoff filters.
}
\examples{
# Linear Regression
set.seed(2024)
p = 100; n = 80; k = 10; scale = 3
Ac = 1:k
rho = 0.3; SigmaX <- toeplitz(rho^(0:(p-1)))
SigmaXhalf = chol(SigmaX)
beta = matrix(0, p, 1)
beta[Ac] = sample(c(-1, 1) * scale, k, replace = TRUE)
X = matrix(rnorm(n * p), n) \%*\% SigmaXhalf
y = X \%*\% beta + rnorm(n)
Xk = create.knockoff(X = X, type = 'shrink', n_ko = 10)
res1 = knockoff.filter(X, y, Xk, statistic = stat.glmnet_coefdiff,
                       aggregate = agg_Freq,
                       offset = 1, fdr = 0.1)
agg_Avg(res1$Ws)

}
\seealso{
Other aggregate: 
\code{\link{agg_BH}()},
\code{\link{agg_Freq}()}
}
\concept{aggregate}
