% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.gaussian.R
\name{create.gaussian}
\alias{create.gaussian}
\title{Model-X Gaussian knockoffs}
\usage{
create.gaussian(X, mu, Sigma, method = c("asdp", "sdp", "equi"), diag_s = NULL)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{mu}{vector of length p, indicating the mean parameter of the Gaussian model for \eqn{X}.}

\item{Sigma}{p-by-p covariance matrix for the Gaussian model of \eqn{X}.}

\item{method}{either "equi", "sdp" or "asdp" (default: "asdp").
This determines the method that will be used to minimize the correlation between the original variables and the knockoffs.}

\item{diag_s}{vector of length p, containing the pre-computed covariances between the original
variables and the knockoffs. This will be computed according to \code{method}, if not supplied.}
}
\value{
A n-by-p matrix of knockoff variables.
}
\description{
Samples multivariate Gaussian model-X knockoff variables.
}
\references{
Candes et al., Panning for Gold: Model-free Knockoffs for High-dimensional Controlled Variable Selection,
arXiv:1610.02351 (2016).
\url{https://web.stanford.edu/group/candes/knockoffs/index.html}
}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.knockoff}()},
\code{\link{create.pc.knockoff}()},
\code{\link{create.pls.knockoff}()},
\code{\link{create.second_order}()},
\code{\link{create.zpls.knockoff}()}
}
\concept{create}
