% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.knockoff.R
\name{create.knockoff}
\alias{create.knockoff}
\title{Generate Knockoff Matrix}
\usage{
create.knockoff(X, type, num, num.comp = 5, verbose = FALSE)
}
\arguments{
\item{X}{An input original design matrix.}

\item{type}{The knockoff type to be generated.
Available options for \code{type} are:
\enumerate{
\item "shrink" for the shrink Gaussian knockoff;
\item "sparse" for the sparse Gaussian knockoff;
\item "pc" for the principal component (PC) knockoff;
\item "pls" for the partial least squares (PLS) knockoff;
\item "zpls" for the GZ's sparse partial least squares knockoff.
}}

\item{num}{The number of knockoff matrices to be created.}

\item{num.comp}{The number of principal components to be used for generating knockoff matrices. Default is 10.}

\item{verbose}{Logical; if TRUE, messages about the progress are printed. Default is FALSE.}
}
\value{
A list of created knockoff matrices.
}
\description{
Generate different types of knockoff matrices given an original one.
}
\examples{
set.seed(10)
X <- matrix(rnorm(100), nrow = 10)
Xk <- create.knockoff(X = X, type = "shrink", num = 5)

}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pc.knockoff}()},
\code{\link{create.second_order}()}
}
\concept{create}
