% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.pc.R
\name{create.pc}
\alias{create.pc}
\title{Create PC Knockoffs}
\usage{
create.pc(X, n_ko = 1, ncomp, verbose = FALSE)
}
\arguments{
\item{X}{A numeric matrix representing the original design matrix.}

\item{n_ko}{An integer specifying the number of knockoff matrices to generate. Default is 1.}

\item{ncomp}{The number of principal components to use in the knockoff generation process.}

\item{verbose}{Logical. Whether to display progress information during knockoff generation. Default is TRUE.}
}
\value{
A list of principal component knockoff matrices. Each matrix corresponds to a generated knockoff.
}
\description{
A sequential algorithm to create non-parametric knockoffs based on principal component regression and residual permutation.
}
\details{
For each original variable \eqn{\mathbf{x}_j}, where \eqn{j = 1, \ldots, p}, the following steps are performed to generate knockoff variables:
\enumerate{
\item Conduct PCA on the matrix \eqn{\left(\mathbf{X}_{-j}, \mathbf{Z}_{1:j-1}\right)}.
\item For a fixed \eqn{K}, fit \eqn{\mathbf{x}_j} on \eqn{K} principal components (PCs). The choice of \eqn{K} involves a tradeoff: a larger \eqn{K} makes the knockoff more similar to the original variable, leading to a lower type I error but weaker power of the test.
\item Compute the residual vector \eqn{\varepsilon_n = \mathbf{x}_j - \hat{\mathbf{x}}_j}.
\item Permute \eqn{\varepsilon_n} randomly and denote the permuted residuals as \eqn{\varepsilon_n^*}.
\item Set \eqn{\mathbf{z}_j = \hat{\mathbf{x}}_j + \varepsilon_n^*} and combine it with the current knockoff matrix \eqn{\mathbf{Z}_{1:j-1}}.
}
}
\examples{
X <- matrix(rnorm(100), nrow = 10)
Xk <- create.pc(X = X, ncomp = 5)

}
\references{
Jiang, Tao, Yuanyuan Li, and Alison A. Motsinger-Reif. "Knockoff boosted tree for model-free variable selection." Bioinformatics 37.7 (2021): 976-983.

Shen, A. et al. (2019). "False discovery rate control in cancer biomarker selection using knockoffs." Cancers, 11, 744.
}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pls}()},
\code{\link{create.second_order}()},
\code{\link{create.seq}()},
\code{\link{create.shrink_Gaussian}()},
\code{\link{create.sparse_Gaussian}()},
\code{\link{create.sparse_seq}()},
\code{\link{create.zpls}()}
}
\concept{create}
