% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.pls.R
\name{create.pls}
\alias{create.pls}
\title{Generate knockoff variables using PLS regression (PLSKO)}
\usage{
create.pls(
  X,
  n_ko = 1,
  ncomp = NULL,
  sparsity = 1,
  nb.list = NULL,
  threshold.abs = NULL,
  threshold.q = 0.9,
  verbose = FALSE
)
}
\arguments{
\item{X}{A numeric matrix or data frame. The original design data matrix with \eqn{n} observations as rows and \eqn{p} variables as columns.}

\item{n_ko}{An integer specifying the number of knockoff variables to generate. Default is 1.}

\item{ncomp}{Optional. An integer specifying the number of components to use in the PLS regression. Default is \code{NULL}, in which case the number of components is chosen empirically.}

\item{sparsity}{Optional. A numeric value between 0 and 1 specifying the sparsity level in the PLS regression. Default is 1 (no sparsity).}

\item{nb.list}{Optional. A list of length \eqn{p} or an adjacency matrix of \eqn{p \times p} that defines the neighbor relationships among variables.
\itemize{
\item A list of length \eqn{p} should include the neighbors' indices of each variable from \eqn{X_1} to \eqn{X_p} in order. The \eqn{i^{th}} element in the list includes the indices of the neighbor variables of \eqn{X_i}, or \code{NULL} when no neighbors.
\item An adjacency matrix should be symmetric with binary elements. \eqn{M_{ij} = 1} indicates that \eqn{X_i} and \eqn{X_j} are neighbors; \eqn{M_{ij} = 0} indicates no neighbor relationship or diagonal entries.
\item If not provided or \code{NULL}, neighborhoods are determined based on correlations.
}}

\item{threshold.abs}{Optional. A value between \eqn{0} and \eqn{1} to specify an absolute correlation threshold for defining neighborhoods.}

\item{threshold.q}{Optional. A numeric value between 0 and 1 indicating the quantile of the correlation values to use as a threshold for defining neighborhoods. Default is 0.9.}

\item{verbose}{Logical. Whether to display progress information during the knockoff generation. Default is TRUE.}
}
\value{
A list of generated knockoff matrices, where each matrix has \eqn{n} rows (observations) and \eqn{p} columns (variables).
}
\description{
This function generates knockoff variables using Partial Least Squares (PLS) regression, following the PLSKO algorithm. It is useful for generating knockoff variables for high-dimensional data.
}
\examples{
set.seed(10)
X <- matrix(rnorm(100), nrow = 10)
Xk <- create.pls(X = X, ncomp = 3)

}
\references{
Yang, Guannan, et al. "PLSKO: a robust knockoff generator to control false discovery rate in omics variable selection." bioRxiv (2024): 2024-08.
}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pc}()},
\code{\link{create.second_order}()},
\code{\link{create.seq}()},
\code{\link{create.shrink_Gaussian}()},
\code{\link{create.sparse_Gaussian}()},
\code{\link{create.sparse_seq}()},
\code{\link{create.zpls}()}
}
\concept{create}
