% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.shrink_Gaussian.R
\name{create.shrink_Gaussian}
\alias{create.shrink_Gaussian}
\title{Create Shrink Gaussian Knockoffs}
\usage{
create.shrink_Gaussian(X, n_ko = 1, verbose = FALSE)
}
\arguments{
\item{X}{A numeric matrix or data frame of predictors for which knockoffs will be generated. Each column represents a variable, and each row corresponds to an observation.}

\item{n_ko}{An integer specifying the number of knockoff matrices to generate. Default is 1.}

\item{verbose}{Logical. If TRUE, displays progress information during knockoff generation. Default is TRUE.}
}
\value{
A list of matrices containing Gaussian knockoff variables corresponding to the original matrix \code{X}. Each matrix in the list has the same dimensions as \code{X}.
}
\description{
This function generates Gaussian knockoff variables using the second-order method with the semi-definite programming (SDP) approach. It is particularly useful for variable selection tasks involving shrinkage methods.
}
\details{
This function uses the semi-definite programming (SDP) method to generate second-order Gaussian knockoffs. It is useful in statistical settings, such as controlling the false discovery rate in variable selection tasks that involve shrinkage methods.
}
\examples{
# Generate a random matrix of predictors
X <- matrix(rnorm(100), nrow = 10, ncol = 10)

# Create shrink Gaussian knockoffs
Xk <- create.shrink_Gaussian(X)

}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pc}()},
\code{\link{create.pls}()},
\code{\link{create.second_order}()},
\code{\link{create.seq}()},
\code{\link{create.sparse_Gaussian}()},
\code{\link{create.sparse_seq}()},
\code{\link{create.zpls}()}
}
\concept{create}
