% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.sparse_Gaussian.R
\name{create.sparse_Gaussian}
\alias{create.sparse_Gaussian}
\title{Create Sparse Gaussian Knockoffs}
\usage{
create.sparse_Gaussian(X, n_ko = 1, verbose = FALSE)
}
\arguments{
\item{X}{A numeric matrix of predictors for which knockoffs will be generated. Each column represents a variable, and each row corresponds to an observation.}

\item{n_ko}{An integer specifying the number of knockoff matrices to generate. Default is 1.}

\item{verbose}{Logical. Whether to display progress information during the knockoff generation. Default is TRUE.}
}
\value{
A list of matrices containing sparse Gaussian knockoff variables corresponding to the original matrix \code{X}. Each matrix in the list has the same dimensions as \code{X}.
}
\description{
Generates sparse Gaussian knockoff variables using the second-order method with the semi-definite programming (SDP) approach.
}
\details{
This function generates knockoff variables using the semi-definite programming (SDP) method for second-order knockoffs with a sparse covariance matrix. Sparsity is controlled through the \code{spcov} package, which allows for shrinkage methods in variable selection tasks.
}
\examples{
X <- matrix(rnorm(100), nrow = 10, ncol = 10)
Xk <- create.sparse_Gaussian(X)

}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pc}()},
\code{\link{create.pls}()},
\code{\link{create.second_order}()},
\code{\link{create.seq}()},
\code{\link{create.shrink_Gaussian}()},
\code{\link{create.sparse_seq}()},
\code{\link{create.zpls}()}
}
\concept{create}
