% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.zpls.R
\name{create.zpls}
\alias{create.zpls}
\title{Generate knockoff variables using sparse partial least squares (SPLS) regression}
\usage{
create.zpls(
  X,
  n_ko = 1,
  ncomp = NULL,
  eta = 0,
  nb.list = NULL,
  threshold.abs = NULL,
  threshold.q = 0.9,
  verbose = FALSE
)
}
\arguments{
\item{X}{A numeric matrix or data frame. The design matrix with \eqn{n} observations as rows and \eqn{p} variables as columns.}

\item{n_ko}{Integer. The number of knockoff copies to generate. Default is 1.}

\item{ncomp}{Optional. Integer specifying the number of components to use in the SPLS regression. Default is 2.}

\item{eta}{Optional. Numeric value between 0 and 1 specifying the sparsity level in the SPLS regression. Default is 0 (no sparsity).}

\item{nb.list}{Optional. A list of length \eqn{p} or a \eqn{p \times p} adjacency matrix defining the neighborhoods of variables.
If not provided or \code{NULL}, neighborhoods are determined based on correlations.}

\item{threshold.abs}{Optional. A numeric value between \eqn{0} and \eqn{1} specifying an absolute correlation threshold to define neighborhoods.}

\item{threshold.q}{Optional. A numeric value between \eqn{0} and \eqn{1} indicating the quantile of the correlation values to use as a threshold. Default is 0.9.}

\item{verbose}{Logical. Whether to display progress information during the knockoff generation. Default is TRUE.}
}
\value{
A matrix of generated knockoff variables of dimensions \eqn{n \times p}.
}
\description{
Generate knockoff variables using sparse partial least squares (SPLS) regression
}
\details{
Knockoff variables are generated by fitting an SPLS regression model for each variable based on its neighborhood.

\strong{Neighborhood Generation}:
\itemize{
\item If \code{threshold.abs} is given, the absolute correlation threshold is used directly.
\item If \code{threshold.q} is given, neighborhoods are determined based on the quantile of the absolute correlation values.
\item If neither is provided, the function defaults to the 90th percentile of the absolute correlation values.
}

\strong{SPLS Regression}:
The fitted values of each variable \eqn{X_j} are calculated using the \code{spls::spls} function. See \code{\link[spls]{spls}} for more details.
}
\examples{
set.seed(10)
X <- matrix(rnorm(100), nrow = 10)
Xk <- create.zpls(X = X, ncomp = 3, eta = 0.3)

}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pc}()},
\code{\link{create.pls}()},
\code{\link{create.second_order}()},
\code{\link{create.seq}()},
\code{\link{create.shrink_Gaussian}()},
\code{\link{create.sparse_Gaussian}()},
\code{\link{create.sparse_seq}()}
}
\concept{create}
