% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.agg.R
\name{empirical_pval}
\alias{empirical_pval}
\title{Compute Empirical P-values from Test Statistics}
\usage{
empirical_pval(test_score, offset = 1)
}
\arguments{
\item{test_score}{A numeric vector of test statistics.}

\item{offset}{A numeric value to adjust p-values. Options are \code{0} or \code{1}. Default is \code{1}.}
}
\value{
A numeric vector of empirical p-values.
}
\description{
Computes empirical p-values from test statistics. Function Adapted from P Tian et al. (2022).
}
\references{
\itemize{
\item Nguyen TB, Chevalier JA, Thirion B, Arlot S. Aggregation of multiple knockoffs. In: International Conference on Machine Learning. PMLR; 2020. p. 7283–93.
\item Tian P, Hu Y, Liu Z et al. Grace-AKO: a novel and stable knockoff filter for variable selection incorporating gene network structures. BMC Bioinformatics 2022;23:478.
}
}
\keyword{internal}
