% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.recovery.R
\name{glmnet.recovery}
\alias{glmnet.recovery}
\title{Simulate from glmnet penalized regression model}
\usage{
glmnet.recovery(y, X, ...)
}
\arguments{
\item{y}{response vector (either "numeric" or "factor") that gets passed to cv.glmnet}

\item{X}{data.frame of covariates that are passed to cv.glmnet}

\item{...}{other parameters passed to the function cv.glmnet}
}
\value{
simulated response
}
\description{
Simulate from glmnet penalized regression model
}
\examples{
set.seed(1)

X = data.frame(matrix(rnorm(100 * 20), 100, 20))
y = X[,1] + rnorm(100)

# simulate from elastic-net regression:
ysim = glmnet.recovery(y=y, X=X)
}
\seealso{
Other recovery: 
\code{\link{ols.recovery}()},
\code{\link{pls.recovery}()},
\code{\link{simple.recovery}()},
\code{\link{spls.recovery}()}
}
\concept{recovery}
