% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filter.R
\name{knockoff.select}
\alias{knockoff.select}
\title{Select Variables based on knockoff statistics}
\usage{
knockoff.select(Ws, fdr = 0.1, offset = 1)
}
\arguments{
\item{Ws}{the test statistics, it is either a nrep by p matrix, or list of length to be nrep, or a vector with length being p.}

\item{fdr}{target false discovery rate (default: 0.1)}

\item{offset}{either 0 or 1 (default: 1). The offset used to compute the rejection threshold on the
statistics. The value 1 yields a slightly more conservative procedure ("knockoffs+") that
controls the FDR according to the usual definition, while an offset of 0 controls a modified FDR.}
}
\value{
An object of class "knockoff.select. This object is a list
containing at least the following components:
\item{W}{computed test statistics}
\item{Ws}{matrix of computed test statistics}
\item{thre}{computed selection threshold}
\item{index}{index of selected variables}
}
\description{
Select Variables based on knockoff statistics
}
