% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_eval.R
\name{perf_eval}
\alias{perf_eval}
\title{Evaluate False Discovery Proportion (FDP) and True Positive Proportion (TPP)}
\usage{
perf_eval(selected, positives, negatives)
}
\arguments{
\item{selected}{A vector of indices of selected variables.}

\item{positives}{A vector of indices of known non-null variables.}

\item{negatives}{A vector of indices of known null variables.}
}
\value{
A numeric vector of length 2 containing:
- `tpr`: The true positive rate.
- `fdr`: The false discovery rate.
}
\description{
This function calculates the false discovery rate and true positive rate
based on selected variables, known null variables (negatives), and known
non-null variables (positives).
}
\examples{
perf_eval(selected = c(1, 2, 3, 5), positives = 1:4, negatives = 5:10)
}
