% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff.threshold.R
\name{print.knockoff.filter}
\alias{print.knockoff.filter}
\title{Print results for the multiple knockoff filter}
\usage{
\method{print}{knockoff.filter}(x, ...)
}
\arguments{
\item{x}{the output of a call to knockoff.filter}

\item{...}{unused}
}
\description{
Prints the list of variables selected by the knockoff filter and the corresponding function call.
}
