% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.Aggregate.R
\name{quantile_aggregation}
\alias{quantile_aggregation}
\title{Quantile Aggregation}
\usage{
quantile_aggregation(pvals, gamma = 0.3)
}
\arguments{
\item{pvals}{A numeric vector of p-values.}

\item{gamma}{A numeric value for the quantile to use in aggregation. Default is \code{0.3}.}
}
\value{
A numeric value representing the aggregated p-value.
}
\description{
Aggregates multiple p-values using quantile aggregation based on Meinshausen et al. (2009). Function Adapted from P Tian et al. (2022).
}
\references{
\itemize{
\item Nguyen TB, Chevalier JA, Thirion B, Arlot S. Aggregation of multiple knockoffs. In: International Conference on Machine Learning. PMLR; 2020. p. 7283–93.
\item Tian P, Hu Y, Liu Z et al. Grace-AKO: a novel and stable knockoff filter for variable selection incorporating gene network structures. BMC Bioinformatics#' @keywords internal
}
}
\keyword{internal}
