% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.recovery.R
\name{simple.recovery}
\alias{simple.recovery}
\title{Simple knockoff generator}
\usage{
simple.recovery(y, X)
}
\arguments{
\item{y}{A response vector that can be either "numeric" or "factor". The vector is passed to \code{cv.glmnet} for fitting the model.}

\item{X}{A data frame of covariates that are passed to \code{cv.glmnet} for model fitting.}
}
\value{
A simulated response vector \code{y.sim} with the same type and length as \code{y}.
}
\description{
This function generates knockoff variables using the following approaches:
\itemize{
\item Continuous variables: Ordinary least squares (OLS) regression.
\item Factor variables: Multinomial logistic regression.
If \code{X} is empty, knockoffs are sampled from the marginal distribution of \code{y}.
}
}
\seealso{
Other recovery: 
\code{\link{glmnet.recovery}()},
\code{\link{ols.recovery}()},
\code{\link{pls.recovery}()},
\code{\link{spls.recovery}()}
}
\concept{recovery}
