% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-spls.knockoff.R
\name{spls.recovery.generator}
\alias{spls.recovery.generator}
\title{Calculate \eqn{\hat{X}} by fitting spls}
\usage{
spls.recovery.generator(Y, X, ncomp, eta)
}
\description{
Calculate \eqn{\hat{X}} by fitting spls
}
\keyword{internal}
