% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.recovery.R
\name{spls.recovery}
\alias{spls.recovery}
\title{Calculate \eqn{\hat{X}} by fitting sparse PLS}
\usage{
spls.recovery(Y, X, ncomp, eta)
}
\arguments{
\item{Y}{A numeric vector representing the response variable.}

\item{X}{A numeric matrix or data frame representing the predictor variables.}

\item{ncomp}{An integer specifying the number of components to include in the sPLS model.}

\item{eta}{A numeric value between 0 and 1 that controls the sparsity of the PLS loadings.}
}
\value{
A numeric vector \eqn{\hat{Y}} containing the fitted values from the sparse PLS model.
}
\description{
This function fits a sparse partial least squares (sPLS) model to the response variable \eqn{Y} using the predictor matrix \eqn{X}. The fitted values \eqn{\hat{Y}} are returned.
}
\details{
This function fits a sparse PLS model using the \code{spls} package. The sparsity level is controlled by the \code{eta} parameter, where higher values lead to sparser loadings. The number of components is specified by \code{ncomp}. The function returns the fitted values of \eqn{Y} from the model.
}
\seealso{
Other recovery: 
\code{\link{glmnet.recovery}()},
\code{\link{ols.recovery}()},
\code{\link{pls.recovery}()},
\code{\link{simple.recovery}()}
}
\concept{recovery}
