% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.SHAP.R
\name{stat.SHAP}
\alias{stat.SHAP}
\title{Importance statistics based on XGBoost}
\usage{
stat.SHAP(X, X_k, y, nrounds = 2, ...)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{X_k}{n-by-p matrix of knockoff variables.}

\item{y}{vector of length n, containing the response variables. If a factor, classification is assumed,
otherwise regression is assumed.}

\item{nrounds}{Number of boosting rounds for training the XGBoost model. Default is 2.}

\item{...}{additional arguments specific to \code{xgboost} (see Details).}
}
\value{
A vector of statistics \eqn{W} of length p.
}
\description{
Computes the difference statistic
\deqn{W_j = |Z_j| - |\tilde{Z}_j|}
where \eqn{Z_j} and \eqn{\tilde{Z}_j} are the SHAP (SHapley Additive exPlanations)
of the jth variable and its knockoff, respectively.
}
\details{
\itemize{
\item In XGBoost, SHAP (SHapley Additive exPlanations) values provide a way to interpret the model's predictions by breaking down the contribution of each feature to the final prediction. SHAP values show how much each feature increases or decreases the prediction compared to the average.
\item XGBoost uses the Tree SHAP algorithm, which efficiently computes these values for tree-based models. This helps in understanding both global feature importance (how features influence the model overall) and local explanations (how features impact individual predictions).
\item Key benefits include transparency, detailed feature importance, and the ability to explain complex models in a clear, interpretable way.
\item saabas is a vector of an individualized heuristic feature attribution method, which can be considered as an approximation for SHAP.
}
}
\examples{
set.seed(2024)
n=80; p=100; k=10; Ac = 1:k; Ic = (k+1):p
X = generate_X(n=n,p=p)
y <- generate_y(X, p_nn=k, a=3)
Xk = create.shrink_Gaussian(X = X, n_ko = 10)
res1 = knockoff.filter(X, y, Xk, statistic = stat.SHAP,
                       offset = 1, fdr = 0.1)
res1
perf_eval(res1$shat,Ac,Ic)

}
\seealso{
Other statistics: 
\code{\link{stat.forward_selection}()},
\code{\link{stat.glmnet_coefdiff}()},
\code{\link{stat.glmnet_lambdadiff}()},
\code{\link{stat.random_forest}()},
\code{\link{stat.sqrt_lasso}()},
\code{\link{stat.stability_selection}()},
\code{\link{stat.xgboost}()}
}
\concept{statistics}
