% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.xgboost.R
\name{stat.xgboost}
\alias{stat.xgboost}
\title{Importance statistics based on xgboost}
\usage{
stat.xgboost(X, X_k, y, family = "gaussian", nrounds = 10, ...)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{X_k}{n-by-p matrix of knockoff variables.}

\item{y}{vector of length n, containing the response variables. If a factor, classification is assumed,
otherwise regression is assumed.}

\item{family}{specifies the type of model to be fit: 'gaussian' for regression or 'binomial' for classification.}

\item{nrounds}{number of boosting rounds for xgboost.}

\item{...}{additional arguments specific to xgboost (see Details).}
}
\value{
A vector of statistics \eqn{W} of length p.
}
\description{
Computes the difference statistic
\deqn{W_j = |Z_j| - |\tilde{Z}_j|}
where \eqn{Z_j} and \eqn{\tilde{Z}_j} are the xgboost feature importances
of the jth variable and its knockoff, respectively.
}
\examples{
set.seed(2024)
n=80; p=100; k=10; Ac = 1:k; Ic = (k+1):p
X = generate_X(n=n,p=p)
y <- generate_y(X, p_nn=k, a=3)
Xk = create.shrink_Gaussian(X = X, n_ko = 10)
res1 = knockoff.filter(X, y, Xk, statistic = stat.xgboost,
                       offset = 1, fdr = 0.1)
res1
perf_eval(res1$shat,Ac,Ic)

}
\seealso{
Other statistics: 
\code{\link{stat.SHAP}()},
\code{\link{stat.forward_selection}()},
\code{\link{stat.glmnet_coefdiff}()},
\code{\link{stat.glmnet_lambdadiff}()},
\code{\link{stat.random_forest}()},
\code{\link{stat.sqrt_lasso}()},
\code{\link{stat.stability_selection}()}
}
\concept{statistics}
