% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_Avg.R
\name{agg_Avg}
\alias{agg_Avg}
\title{Aggregated Knockoff with Average Test Statistics}
\usage{
agg_Avg(Ws_mat, fdr = 0.05, offset = 0)
}
\arguments{
\item{Ws_mat}{A matrix where each row represents the test statistics from a different knockoff filter and each column corresponds to a variable.}

\item{fdr}{A numeric value specifying the target false discovery rate (FDR) level. Default is \eqn{0.05}.}

\item{offset}{An integer (0 or 1) specifying the offset in the threshold calculation. Default is \eqn{0} (liberal control).}
}
\value{
A vector \code{shat} containing the indices of selected variables after averaging the test statistics and applying the knockoff filter.
}
\description{
Performs Aggregated Knockoff analysis by averaging the test statistics across multiple knockoff filters.
}
\examples{
set.seed(2024)
p = 100; n = 80
X = generate_X(n=80,p=100)
y <- generate_y(X, p_nn=10, a=3)
Xk = create.shrink_Gaussian(X = X, n_ko = 10)
res1 = knockoff.filter(X, y, Xk, statistic = stat.glmnet_coefdiff,
                       aggregate = agg_Freq,
                       offset = 1, fdr = 0.1)
res1
agg_Avg(res1$Ws)

}
\seealso{
Other aggregate: 
\code{\link{agg_BH}()},
\code{\link{agg_Freq}()}
}
\concept{aggregate}
