% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.knockoff.R
\name{create.knockoff}
\alias{create.knockoff}
\title{Generate Knockoff Matrix}
\usage{
create.knockoff(X, type, num, num.comp = 10, verbose = FALSE)
}
\arguments{
\item{X}{An input original design matrix.}

\item{type}{The knockoff type to be generated. There are three choices available: (1) "shrink" for the shrink Gaussian knockoff;
(2) "sparse" for the sparse Gaussian knockoff; and (3) "pc" for the pricial component knockoff.}

\item{num}{The number of knockoff matrices to be created.}

\item{num.comp}{The number of pricial components to be used for generating knockoff matrices, the default is 10.}
}
\value{
A list of created knockoff matrices.
}
\description{
Generate different types of knockoff matrices given an original one.
}
\examples{
set.seed(10)
X <- matrix(rnorm(100), nrow = 10)
Xk <- create.knockoff(X = X, type = "shrink", num = 5)

}
\references{

}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pc.knockoff}()},
\code{\link{create.second_order}()}
}
\concept{create}
