% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.knockoff.R
\name{create.knockoff}
\alias{create.knockoff}
\title{Generate Knockoff Matrix}
\usage{
create.knockoff(X, type, num, ncomp = 10, verbose = FALSE, ...)
}
\arguments{
\item{X}{A numeric matrix representing the original design matrix.}

\item{type}{The knockoff type to be generated. Available options for \code{type} are:
\enumerate{
\item "shrink" for the shrink Gaussian knockoff;
\item "sparse" for the sparse Gaussian knockoff;
\item "pc" for the principal component (PC) knockoff;
\item "pls" for the partial least squares (PLS) knockoff;
\item "zpls" for the GZ's sparse partial least squares knockoff.
}}

\item{num}{A positive integer indicating the number of knockoff matrices to be created.}

\item{ncomp}{The number of principal components to be used for generating knockoff matrices. Default is 10.}

\item{verbose}{Logical; if TRUE, messages about the progress are printed. Default is FALSE.}

\item{...}{Additional arguments to be passed to specific knockoff creation functions.}
}
\value{
A list of created knockoff matrices. Each element of the list corresponds to a generated knockoff matrix.
}
\description{
Generate different types of knockoff matrices given an original one.
}
\examples{
set.seed(10)
X <- matrix(rnorm(100), nrow = 10)
Xk <- create.knockoff(X = X, type = "shrink", num = 5)

}
\seealso{
Other create: 
\code{\link{create.fixed}()},
\code{\link{create.gaussian}()},
\code{\link{create.pc.knockoff}()},
\code{\link{create.pls.knockoff}()},
\code{\link{create.second_order}()},
\code{\link{create.zpls.knockoff}()}
}
\concept{create}
