% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filter.R
\name{knockoff.filter}
\alias{knockoff.filter}
\title{The Knockoff Filter}
\usage{
knockoff.filter(
  X,
  y,
  Xk = NULL,
  statistic = stat.glmnet_coefdiff,
  aggregate = agg_Freq,
  fdr = 0.1,
  offset = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{X}{A numeric n-by-p matrix or data frame of predictors.}

\item{y}{A response vector of length n.}

\item{Xk}{A list of Knockoff copys.}

\item{statistic}{A function to compute test statistics (default: \code{stat.glmnet_coefdiff}).}

\item{aggregate}{Function to aggregate results from multiple knockoffs (default: \code{agg_Freq}).}

\item{fdr}{Target false discovery rate (default: 0.1).}

\item{offset}{Offset for threshold computation (0 or 1; default: 1).}

\item{verbose}{Logical; if TRUE, prints progress messages during knockoff generation and statistic calculation (default: FALSE).}

\item{...}{Additional arguments passed to the \code{statistic} function.}
}
\value{
An object of class \code{knockoff.filter}, containing:
\itemize{
\item Single Knockoff Case:
\itemize{
\item \strong{call}: The matched call of the function.
\item \strong{W}: The test statistics for the original variables.
\item \strong{threshold}: The computed selection threshold.
\item \strong{shat}: The indices of variables selected based on the threshold.
}
\item Multiple Knockoff Case:
\itemize{
\item \strong{call}: The matched call of the function.
\item \strong{shat}: The aggregated indices of selected variables.
\item \strong{Ws}: The matrix of test statistics for multiple knockoff copies.
\item \strong{thresholds}: A vector of thresholds for each knockoff.
\item \strong{shat_list}: A list where each element contains the indices of selected variables for a corresponding knockoff copy.
\item \strong{shat_mat}: A binary matrix where each row indicates the selected variables for a specific knockoff copy (1 for selected, 0 for not selected).
}
}
}
\description{
This function runs the Knockoff procedure, selecting variables relevant for predicting
the outcome of interest using knockoffs and test statistics.
}
\examples{
# Linear Regression
set.seed(2024)
n=80; p=100; k=10; Ac = 1:k; Ic = (k+1):p
X = generate_X(n=n,p=p)
y <- generate_y(X, p_nn=k, a=3)
Xk = create.shrink_Gaussian(X = X, n_ko = 10)
res1 = knockoff.filter(X, y, Xk, statistic = stat.glmnet_coefdiff,
                       offset = 1, fdr = 0.1)
res1
perf_eval(res1$shat,Ac,Ic)

# Logistic Regression
lp <- generate_lp(X, p_nn=k, a=3)
pis <- plogis(lp)
Y <- factor(rbinom(n, 1, pis))
res2 = knockoff.filter(X, Y, Xk, statistic = stat.glmnet_coefdiff,
                       family = 'binomial', offset = 0, fdr = 0.2)
res2
perf_eval(res2$shat,Ac,Ic)

}
\references{
Candes, E., Fan, Y., Janson, L., & Lv, J. (2018). Panning for gold: ‘model-X’ knockoffs for high dimensional controlled variable selection. Journal of the Royal Statistical Society Series B: Statistical Methodology, 80(3), 551-577.
}
