% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.xgboost.R
\name{stat.xgboost}
\alias{stat.xgboost}
\title{Importance statistics based on xgboost}
\usage{
stat.xgboost(X, X_k, y, family = "gaussian", nrounds = 10, ...)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{X_k}{n-by-p matrix of knockoff variables.}

\item{y}{vector of length n, containing the response variables. If a factor, classification is assumed,
otherwise regression is assumed.}

\item{family}{specifies the type of model to be fit: 'gaussian' for regression or 'binomial' for classification.}

\item{nrounds}{number of boosting rounds for xgboost.}

\item{...}{additional arguments specific to xgboost (see Details).}
}
\value{
A vector of statistics \eqn{W} of length p.
}
\description{
Computes the difference statistic
  \deqn{W_j = |Z_j| - |\tilde{Z}_j|}
where \eqn{Z_j} and \eqn{\tilde{Z}_j} are the xgboost feature importances
of the jth variable and its knockoff, respectively.
}
