% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.xgboost.R
\name{stat.xgboost}
\alias{stat.xgboost}
\title{Importance statistics based on xgboost}
\usage{
stat.xgboost(X, X_k, y, family = "gaussian", nrounds = 10)
}
\arguments{
\item{X}{n-by-p matrix of original variables.}

\item{X_k}{n-by-p matrix of knockoff variables.}

\item{y}{vector of length n, containing the response variables. If a factor, classification is assumed,
otherwise regression is assumed.}

\item{...}{additional arguments specific to \code{ranger} (see Details).}
}
\value{
A vector of statistics \eqn{W} of length p.
}
\description{
Computes the difference statistic
\deqn{W_j = |Z_j| - |\tilde{Z}_j|}
where \eqn{Z_j} and \eqn{\tilde{Z}_j} are the xgboost feature importances
of the jth variable and its knockoff, respectively.
}
\details{
.
}
\examples{
# Synthetic Data
set.seed(2024)
p=200; n=100; k=15
mu = rep(0,p); Sigma = diag(p)
X = matrix(rnorm(n*p),n)
nonzero = 1:k
beta = 3.5 * (1:p \%in\% nonzero)
y = X \%*\% beta + rnorm(n)

# Knockoff Procedure
Xk = create.knockoff(X = X, type = 'shrink', num = 2)
res = knockoff.filter(X,y,Xk,statistic = stat.xgboost,family='gaussian')
res$shat

}
\seealso{
Other statistics: 
\code{\link{stat.SHAP}()},
\code{\link{stat.forward_selection}()},
\code{\link{stat.glmnet_coefdiff}()},
\code{\link{stat.glmnet_lambdadiff}()},
\code{\link{stat.random_forest}()},
\code{\link{stat.sqrt_lasso}()},
\code{\link{stat.stability_selection}()}
}
\concept{statistics}
